Better organization.  Some progress towards switching between Uniscribe and a more simple layout process.

+++ TextStyle.h
@@ -33,9 +33,9 @@ public:
 	int FontFallback( HDC, int style, UTF16Ref );
 	bool HasMissingGlyphs( int style, ArrayOf<WORD> glyphs ) const;
 
-	LONG lineHeight;
-	LONG avgCharWidth;
-	LONG tabSize;
+	int lineHeight;
+	int avgCharWidth;
+	int tabSize;
 
 	HBRUSH gutterBrush;
 	HBRUSH selectionBrush;
